psh: touch: move rootfs cases to separate test
JIRA: CI-154
